Rice-Model-Detection/Rice-Model-Detection/requirements.txt
Flask>=2.3.0
tensorflow>=2.13.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.19.0
pandas>=1.3.0
gunicorn>=20.1.0
